renamed package name to add more abstraction
