+++ pom.xml
@@ -5,10 +5,10 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>org.example</groupId>
-  <artifactId>Factorial2</artifactId>
+  <artifactId>Math Utils</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>Factorial2</name>
+  <name>Math Utils</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
